requirements.txt
beautifulsoup4
environs
botocore
aiobotocore
python-dotenv
